static collision finish
box and sphere can properly collide with each other and objects of same type as well as planes. with this the static collisions are pretty much done for now.
also added comparign operators for vector3f to test if a vector is greater or less than in all three dimensions.